minor refactoring..
