+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/techknowlogic/util/TeamShippingElementDetector.java
@@ -1,10 +1,11 @@
-package org.firstinspires.ftc.teamcode.techknowlogic.util.scanner;
+package org.firstinspires.ftc.teamcode.techknowlogic.util;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.teamcode.techknowlogic.util.RobotPosition;
+import org.firstinspires.ftc.teamcode.techknowlogic.util.ScannerCoordinates;
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
 import org.opencv.core.Point;
